third_party/WebKit/LayoutTests/media/video-controls-touch-events-captured-expected.txt

This tests that touch events on the controls will not be seen by the video element.

EVENT(loadeddata)
END OF TEST
